sql: DELETEME - attempt to make sqlparser.execute reentrable
This is mostly failed attempt to implement properly scenario
when previously parsed AST has been reused for vdbe regeneration.

When query is without errors - it's mostly working, but failing
miserably when we need to raise error due to some comumn mismatch
for compount select queries.

Eventually we will return to this code, but for will plan to disable
caching right away.