sql: disable ast parsing caching
**Extremely awkward - should be cleaned up someday**

As a work-around for problem of improper 2nd and onward
time call for vdbe generation over the previously parsed
AST datastructures, we rework the way AST kept and dealt
with: we do not put ast to prepare cache, but rathe return
cdata pointer directly, without caching.
It's working for SELECT, but not work for DDL, and other kinds
DML/DQL. Will be addressed the next commit.